services/backend/requirements.txt
ultralytics==8.0.227
opencv-python
paho-mqtt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
paho-mqtt
